use powershell winget rather than ./winget.exe
Using ./winget.exe threw an error on my clean windows 11 build - but you can just call winget from powershell so I did that and it worked.  I assume you can clear up the code that finds the Winget Path but I don't want to mess with that in case you need it elsewhere.