Update RemoveBloat.ps1
Updates all searches of HKCU to include a check for if the script is running in the SYSTEM context. This allows the script to be run from RMM tools in the SYSTEM context without throwing errors when attempting to search registry keys that don't exist (because the system user doesn't have them).